pyproject.toml
[tool.poetry]
name = "rapidboard-ai"
version = "0.1.0"
description = "A LangGraph based AI agent project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^1.1.6"
langgraph = "^1.0.5"
langsmith = "^0.5.2"
python-dotenv = "^1.2.1"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
pinecone = "^8.0.0"
langchain-community = "^0.4.1"
unstructured = "^0.18.24"
pypdf = "^6.5.0"
python-docx = "^1.2.0"
duckduckgo-search = "^8.1.1"
ddgs = "^9.10.0"
langchain = "^1.2.0"
langchain-text-splitters = "^1.1.0"
python-multipart = "^0.0.21"
google-genai = "^0.3.0"
websockets = "^14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:run"
serve = "src.api.app:run_server"
